feat(Region Area): adds area creation to region handlers
also refactor into mulitple files
